allow specifying `ENUM XXX(` which prepends XXX to each constant name.
This does not change the printing or parsing of enum values.

enum constants reside in the package namespace, so we might need prefixes
to distinguish between values with the same name from different enumerated
types.  Printing and parsing do not need the prefix because these
happen in the context of a known type.

When go-enum is run *without* the --noprefix option, the enum typename is
effectively used as the namespace prefix, but this can be unwieldy, so this
commit provides another option.